requirements.txt
et-xmlfile==1.0.1
Faker==5.4.0
jdcal==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3